WA-268 Update redux reducer to lock-unlock the account
